feat:일반공격 데미지 로직 추가 및 float로 고침

+++ Assets/01.Script/Enemy/Enemy.cs
@@ -41,7 +41,7 @@ public class Enemy : MonoBehaviour, IDamageable
         //OnPlayerStatusChanged?.Invoke();
         //OnPlayerInvChanged?.Invoke(); // 초기 로드 후 UI에 알려줌
     }
-    public void TakePhisicalDamage(int amount)
+    public void TakePhisicalDamage(float amount)
     {
         float hp = CurrentStatus.HP;
         hp -= amount;
